# api/swagger.yaml
openapi: 3.0.1
info:
  title: PrImBoard
  description: >-
    PrImBoard (Private Image Board) can be best described as an image board for
    all the picures and videos you have taken. You can invite users to the board
    and share specific images with them or your family members!
  version: 1.0.0
  contact: {}
servers:
  - url: 'https://localhost/api/v1'
  - url: 'http://localhost/api/v1'
security: []
...
paths:
...
  '/tag/{id}':
    get:
      summary: Get tag by id
      description: ''
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      operationId: getTagById
      responses:
        '200':
          description: Got tag successful
        '400':
          description: Invalid id specified
        '401':
          description: Authentication information is missing or invalid
        '404':
          description: Tag not found
      tags:
        - tag
    put:
      summary: Update tag by id
      description: ''
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      operationId: updateTagBayId
      responses:
        '200':
          description: Updated tag successful
        '400':
          description: Invalid id specified
        '401':
          description: Authentication information is missing or invalid
        '404':
          description: Tag not found
      tags:
        - tag
    delete:
      summary: Delete Tag by id
      description: ''
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      operationId: deleteTagById
      responses:
        '200':
          description: Deleted tag successful
        '400':
          description: Invalid id specified
        '401':
          description: Authentication information is missing or invalid
        '404':
          description: Tag not found
      tags:
        - tag
components:
...
  securitySchemes: {}
...